improve select performance
there are two types for elements added to compare array. suppose
current smallest sum = A[i] + B[j], next element needed to added to
compare array are:
1) (A[i+1],B[j])
2) (A[i],B[j+1])
if there is already exist element (A[i+1],B[j-1]) for 1) and
(A[i-1], B[j+1]) for 2) in the compare array, obviously, element `1)`
or `2)` is bigger than elements in the compare array. there is no
need to add this element to compare array.